+++ recipes/tests/test_recipe_search_view.py
@@ -1,0 +1,24 @@
+from django.urls import reverse, resolve
+from recipes import views
+from .test_recipe_base import RecipeTestBase
+
+
+class RecipesSearchViewsTest(RecipeTestBase):
+
+    def test_search_view_uses_correct_view_function(self):
+        resolved = resolve(reverse('recipes:search'))
+        self.assertIs(resolved.func, views.search)
+
+    def test_search_view_returns_correct_template(self):
+        response = self.client.get(reverse('recipes:search') + '?q=recipes')
+        self.assertTemplateUsed(response, 'recipes/pages/search.html')
+
+    def test_recipe_search_raises_404_if_no_search_term(self):
+        url = reverse('recipes:search')  # + '?q=recipes'
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 404)
+
+    def test_recipe_search_term_is_on_page_title_and_scaped(self):
+        url = reverse('recipes:search') + '?q=Test'
+        response = self.client.get(url)
+        self.assertIn('Search for &quot;Test&quot;', response.content.decode('utf-8'))  # noqa: E501
